Move functions related to failure policy to failure_policy.go file.
