+++ pkg/controllers/failure_policy.go
@@ -72,6 +72,58 @@ func executeFailurePolicy(ctx context.Context, js *jobset.JobSet, ownedJobs *chi
 	return nil
 }
 
+// makeFailedConditionOpts returns the options we use to generate the JobSet failed condition.
+func makeFailedConditionOpts(reason, msg string) *conditionOpts {
+	return &conditionOpts{
+		condition: &metav1.Condition{
+			Type:    string(jobset.JobSetFailed),
+			Status:  metav1.ConditionStatus(corev1.ConditionTrue),
+			Reason:  reason,
+			Message: msg,
+		},
+		eventType: corev1.EventTypeWarning,
+	}
+}
+
+// setJobSetFailedCondition sets a condition on the JobSet status indicating it has failed.
+func setJobSetFailedCondition(ctx context.Context, js *jobset.JobSet, reason, msg string, updateStatusOpts *statusUpdateOpts) {
+	setCondition(js, makeFailedConditionOpts(reason, msg), updateStatusOpts)
+}
+
+// findJobFailureTime is a helper function which extracts the Job failure time from a Job,
+// if the JobFailed condition exists and is true.
+func findJobFailureTime(job *batchv1.Job) *metav1.Time {
+	failureCondition := findJobFailureCondition(job)
+	if failureCondition == nil {
+		return nil
+	}
+	return &failureCondition.LastTransitionTime
+}
+
+// findFirstFailedJob accepts a slice of failed Jobs and returns the Job which has a JobFailed condition
+// with the oldest transition time.
+func findFirstFailedJob(failedJobs []*batchv1.Job) *batchv1.Job {
+	var (
+		firstFailedJob   *batchv1.Job
+		firstFailureTime *metav1.Time
+	)
+	for _, job := range failedJobs {
+		failureTime := findJobFailureTime(job)
+		// If job has actually failed and it is the first (or only) failure we've seen,
+		// store the job for output.
+		if failureTime != nil && (firstFailedJob == nil || failureTime.Before(firstFailureTime)) {
+			firstFailedJob = job
+			firstFailureTime = failureTime
+		}
+	}
+	return firstFailedJob
+}
+
+// messageWithFirstFailedJob appends the first failed job to the original event message in human readable way.
+func messageWithFirstFailedJob(msg, firstFailedJobName string) string {
+	return fmt.Sprintf("%s (first failed job: %s)", msg, firstFailedJobName)
+}
+
 // ruleIsApplicable returns true if the failed job and job failure reason match the failure policy rule.
 // The function returns false otherwise.
 func ruleIsApplicable(rule jobset.FailurePolicyRule, failedJob *batchv1.Job, jobFailureReason string) bool {
@@ -90,6 +142,21 @@ func ruleIsApplicable(rule jobset.FailurePolicyRule, failedJob *batchv1.Job, job
 
 	ruleAppliesToParentReplicatedJob := len(rule.TargetReplicatedJobs) == 0 || slices.Contains(rule.TargetReplicatedJobs, parentReplicatedJob)
 	return ruleAppliesToParentReplicatedJob
+}
+
+// findJobFailureTimeAndReason is a helper function which extracts the Job failure condition from a Job,
+// if the JobFailed condition exists and is true.
+func findJobFailureCondition(job *batchv1.Job) *batchv1.JobCondition {
+	if job == nil {
+		return nil
+	}
+	for _, c := range job.Status.Conditions {
+		// If this Job failed before the oldest known Job failiure, update the first failed job.
+		if c.Type == batchv1.JobFailed && c.Status == corev1.ConditionTrue {
+			return &c
+		}
+	}
+	return nil
 }
 
 // findFirstFailedPolicyRuleAndJob returns the first failure policy rule matching a failed child job.
