+++ pkg/controllers/failure_policy_test.go
@@ -19,12 +19,63 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/stretchr/testify/assert"
 	batchv1 "k8s.io/api/batch/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/ptr"
 
 	jobset "sigs.k8s.io/jobset/api/jobset/v1alpha2"
 	testutils "sigs.k8s.io/jobset/pkg/util/testing"
 )
+
+func TestFindFirstFailedJob(t *testing.T) {
+	testCases := []struct {
+		name       string
+		failedJobs []*batchv1.Job
+		expected   *batchv1.Job
+	}{
+		{
+			name:       "No failed jobs",
+			failedJobs: []*batchv1.Job{},
+			expected:   nil,
+		},
+		{
+			name: "Single failed job",
+			failedJobs: []*batchv1.Job{
+				jobWithFailedCondition("job1", time.Now().Add(-1*time.Hour)),
+			},
+			expected: jobWithFailedCondition("job1", time.Now().Add(-1*time.Hour)),
+		},
+		{
+			name: "Multiple failed jobs, earliest first",
+			failedJobs: []*batchv1.Job{
+				jobWithFailedCondition("job1", time.Now().Add(-3*time.Hour)),
+				jobWithFailedCondition("job2", time.Now().Add(-5*time.Hour)),
+			},
+			expected: jobWithFailedCondition("job2", time.Now().Add(-5*time.Hour)),
+		},
+		{
+			name: "Jobs without failed condition",
+			failedJobs: []*batchv1.Job{
+				{ObjectMeta: metav1.ObjectMeta{Name: "job1"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "job2"}},
+			},
+			expected: nil,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			result := findFirstFailedJob(tc.failedJobs)
+			if result != nil && tc.expected != nil {
+				assert.Equal(t, result.Name, tc.expected.Name)
+			} else if result != nil && tc.expected == nil || result == nil && tc.expected != nil {
+				t.Errorf("Expected: %v, got: %v)", result, tc.expected)
+			}
+		})
+	}
+}
 
 func TestFailurePolicyRuleIsApplicable(t *testing.T) {
 	var (
