+++ pkg/controllers/jobset_controller_test.go
@@ -33,7 +33,6 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/stretchr/testify/assert"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -1102,55 +1101,6 @@ func TestCalculateReplicatedJobStatuses(t *testing.T) {
 			}
 			if diff := cmp.Diff(tc.expected, statuses, cmpopts.SortSlices(less)); diff != "" {
 				t.Errorf("calculateReplicatedJobStatuses() mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestFindFirstFailedJob(t *testing.T) {
-	testCases := []struct {
-		name       string
-		failedJobs []*batchv1.Job
-		expected   *batchv1.Job
-	}{
-		{
-			name:       "No failed jobs",
-			failedJobs: []*batchv1.Job{},
-			expected:   nil,
-		},
-		{
-			name: "Single failed job",
-			failedJobs: []*batchv1.Job{
-				jobWithFailedCondition("job1", time.Now().Add(-1*time.Hour)),
-			},
-			expected: jobWithFailedCondition("job1", time.Now().Add(-1*time.Hour)),
-		},
-		{
-			name: "Multiple failed jobs, earliest first",
-			failedJobs: []*batchv1.Job{
-				jobWithFailedCondition("job1", time.Now().Add(-3*time.Hour)),
-				jobWithFailedCondition("job2", time.Now().Add(-5*time.Hour)),
-			},
-			expected: jobWithFailedCondition("job2", time.Now().Add(-5*time.Hour)),
-		},
-		{
-			name: "Jobs without failed condition",
-			failedJobs: []*batchv1.Job{
-				{ObjectMeta: metav1.ObjectMeta{Name: "job1"}},
-				{ObjectMeta: metav1.ObjectMeta{Name: "job2"}},
-			},
-			expected: nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			result := findFirstFailedJob(tc.failedJobs)
-			if result != nil && tc.expected != nil {
-				assert.Equal(t, result.Name, tc.expected.Name)
-			} else if result != nil && tc.expected == nil || result == nil && tc.expected != nil {
-				t.Errorf("Expected: %v, got: %v)", result, tc.expected)
 			}
 		})
 	}
